chore: refactory e2e funcs
