+++ test/e2e/configurer/chain/queries_btcstaking.go
@@ -1,6 +1,7 @@
 package chain
 
 import (
+	"encoding/hex"
 	"fmt"
 	"net/url"
 
@@ -55,6 +56,18 @@ func (n *NodeConfig) QueryFinalityProviders() []*bstypes.FinalityProviderRespons
 	require.NoError(n.t, err)
 
 	return resp.FinalityProviders
+}
+
+func (n *NodeConfig) QueryFinalityProvider(btcPkHex string) *bstypes.FinalityProviderResponse {
+	path := fmt.Sprintf("/babylon/btcstaking/v1/finality_providers/%s/finality_provider", btcPkHex)
+	bz, err := n.QueryGRPCGateway(path, url.Values{})
+	require.NoError(n.t, err)
+
+	var resp bstypes.QueryFinalityProviderResponse
+	err = util.Cdc.UnmarshalJSON(bz, &resp)
+	require.NoError(n.t, err)
+
+	return resp.FinalityProvider
 }
 
 func (n *NodeConfig) QueryActiveFinalityProvidersAtHeight(height uint64) []*ftypes.ActiveFinalityProvidersAtHeightResponse {
@@ -205,4 +218,99 @@ func (n *NodeConfig) QueryFinalityProvidersDelegations(fpsBTCPK ...string) []*bs
 		}
 	}
 	return pendingDelsResp
+}
+
+// ParseRespBTCDelToBTCDel parses an BTC delegation response to BTC Delegation
+func ParseRespBTCDelToBTCDel(resp *bstypes.BTCDelegationResponse) (btcDel *bstypes.BTCDelegation, err error) {
+	stakingTx, err := hex.DecodeString(resp.StakingTxHex)
+	if err != nil {
+		return nil, err
+	}
+
+	delSig, err := bbn.NewBIP340SignatureFromHex(resp.DelegatorSlashSigHex)
+	if err != nil {
+		return nil, err
+	}
+
+	slashingTx, err := bstypes.NewBTCSlashingTxFromHex(resp.SlashingTxHex)
+	if err != nil {
+		return nil, err
+	}
+
+	btcDel = &bstypes.BTCDelegation{
+		StakerAddr:       resp.StakerAddr,
+		BtcPk:            resp.BtcPk,
+		FpBtcPkList:      resp.FpBtcPkList,
+		StartHeight:      resp.StartHeight,
+		StakingTime:      resp.StakingTime,
+		EndHeight:        resp.EndHeight,
+		TotalSat:         resp.TotalSat,
+		StakingTx:        stakingTx,
+		DelegatorSig:     delSig,
+		StakingOutputIdx: resp.StakingOutputIdx,
+		CovenantSigs:     resp.CovenantSigs,
+		UnbondingTime:    resp.UnbondingTime,
+		SlashingTx:       slashingTx,
+	}
+
+	if resp.UndelegationResponse != nil {
+		ud := resp.UndelegationResponse
+		unbondTx, err := hex.DecodeString(ud.UnbondingTxHex)
+		if err != nil {
+			return nil, err
+		}
+
+		slashTx, err := bstypes.NewBTCSlashingTxFromHex(ud.SlashingTxHex)
+		if err != nil {
+			return nil, err
+		}
+
+		delSlashingSig, err := bbn.NewBIP340SignatureFromHex(ud.DelegatorSlashingSigHex)
+		if err != nil {
+			return nil, err
+		}
+
+		btcDel.BtcUndelegation = &bstypes.BTCUndelegation{
+			UnbondingTx:              unbondTx,
+			CovenantUnbondingSigList: ud.CovenantUnbondingSigList,
+			CovenantSlashingSigs:     ud.CovenantSlashingSigs,
+			SlashingTx:               slashTx,
+			DelegatorSlashingSig:     delSlashingSig,
+		}
+
+		if ud.DelegatorUnbondingInfoResponse != nil {
+			var spendStakeTx []byte = make([]byte, 0)
+			if ud.DelegatorUnbondingInfoResponse.SpendStakeTxHex != "" {
+				spendStakeTx, err = hex.DecodeString(ud.DelegatorUnbondingInfoResponse.SpendStakeTxHex)
+				if err != nil {
+					return nil, err
+				}
+			}
+
+			btcDel.BtcUndelegation.DelegatorUnbondingInfo = &bstypes.DelegatorUnbondingInfo{
+				SpendStakeTx: spendStakeTx,
+			}
+		}
+	}
+
+	return btcDel, nil
+}
+
+// ParseRespsBTCDelToBTCDel parses an BTC delegation response to BTC Delegation
+func ParseRespsBTCDelToBTCDel(resp *bstypes.BTCDelegatorDelegationsResponse) (btcDels *bstypes.BTCDelegatorDelegations, err error) {
+	if resp == nil {
+		return nil, nil
+	}
+	btcDels = &bstypes.BTCDelegatorDelegations{
+		Dels: make([]*bstypes.BTCDelegation, len(resp.Dels)),
+	}
+
+	for i, delResp := range resp.Dels {
+		del, err := ParseRespBTCDelToBTCDel(delResp)
+		if err != nil {
+			return nil, err
+		}
+		btcDels.Dels[i] = del
+	}
+	return btcDels, nil
 }
